+++ capitalize_decorator.rb
@@ -1,0 +1,7 @@
+require_relative 'base_decorator'
+
+class CapitilizeDecorator < BaseDecorator
+  def correct_name
+    @nameable.correct_name.capitilize
+  end
+end
